Move `__memcpy_async_barrier` to its own file
