+++ libcudacxx/include/cuda/__memcpy_async/memcpy_completion.h
@@ -24,10 +24,10 @@
 #include <cuda/__barrier/async_contract_fulfillment.h>
 #include <cuda/__barrier/barrier_expect_tx.h>
 #include <cuda/__barrier/barrier_thread_scope_block.h>
-#include <cuda/__barrier/completion_mechanism.h>
-#include <cuda/__barrier/is_local_smem_barrier.h>
-#include <cuda/__barrier/try_get_barrier_handle.h>
 #include <cuda/__fwd/pipeline.h>
+#include <cuda/__memcpy_async/completion_mechanism.h>
+#include <cuda/__memcpy_async/is_local_smem_barrier.h>
+#include <cuda/__memcpy_async/try_get_barrier_handle.h>
 #include <cuda/std/__atomic/scopes.h>
 #include <cuda/std/cstdint>
